Move index to legacy tasks controller
