+++ app/controllers/legacy_tasks_controller.rb
@@ -2,8 +2,30 @@ class LegacyTasksController < ApplicationController
   before_action :verify_queue_access
   before_action :verify_task_assignment_access, only: [:create, :update]
 
+  ROLES = %w[judge attorney].freeze
+
   def set_application
     RequestStore.store[:application] = "queue"
+  end
+
+  def index
+    current_role = params[:role] || current_user.vacols_roles.first
+    return invalid_role_error unless ROLES.include?(current_role)
+    respond_to do |format|
+      format.html do
+        render "queue/show"
+      end
+      format.json do
+        MetricsService.record("VACOLS: Get all tasks with appeals for #{params[:user_id]}",
+                              name: "LegacyTasksController.index") do
+          tasks, appeals = WorkQueue.tasks_with_appeals(user, current_role)
+          render json: {
+            tasks: json_tasks(tasks),
+            appeals: json_appeals(appeals)
+          }
+        end
+      end
+    end
   end
 
   def create
@@ -26,6 +48,11 @@ class LegacyTasksController < ApplicationController
 
   private
 
+  def user
+    @user ||= User.find(params[:user_id])
+  end
+  helper_method :user
+
   def legacy_task_params
     params.require("tasks")
       .permit(:appeal_id)
@@ -37,6 +64,20 @@ class LegacyTasksController < ApplicationController
     ActiveModelSerializers::SerializableResource.new(
       task,
       serializer: ::WorkQueue::TaskSerializer
+    ).as_json
+  end
+
+  def json_appeals(appeals)
+    ActiveModelSerializers::SerializableResource.new(
+      appeals,
+      each_serializer: ::WorkQueue::LegacyAppealSerializer
+    ).as_json
+  end
+
+  def json_tasks(tasks)
+    ActiveModelSerializers::SerializableResource.new(
+      tasks,
+      each_serializer: ::WorkQueue::TaskSerializer
     ).as_json
   end
 end
